% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co.co.R
\name{co.co}
\alias{co.co}
\title{co.co}
\usage{
co.co(
  study,
  group,
  t,
  n,
  mean,
  m,
  s,
  se,
  ll95,
  ul95,
  ll90,
  ul90,
  ll99,
  ul99,
  a,
  b,
  lq,
  uq,
  data = NULL,
  meanSd.meanAB = "range",
  meanSd.mIqrRange = "wanEtal",
  meanSd.mIqr = "wanEtal",
  meanSd.mRange = "wanEtal",
  meanSd.mean = "furukawaEtal"
)
}
\arguments{
\item{study}{A vector with study labels (e.g., "1" for study 1, "2" for study 2, ...).}

\item{group}{A vector with group labels (e.g., "1" for intervention group, "2" for control group).}

\item{t}{A vector with time points of outcome measurement (e.g., "1" for baseline, "2" for postintervention, and "3" for change from baseline).}

\item{n}{A vector with numbers of observations in groups at time points.}

\item{mean}{An optional vector of means of the outcome in groups at time points.}

\item{m}{An optional vector of medians of the outcome in groups at time points.}

\item{s}{An optional vector of standard deviations of the outcome in groups at time points.}

\item{se}{An optional vector of standard errors of the mean of the outcome in groups at time points.}

\item{ll95}{An optional vector of the lower limits of the 95 percent confidence interval of the mean outcome in groups at time points.}

\item{ul95}{An optional vector of the upper limits of the 95 percent confidence interval of the mean outcome in groups at time points.}

\item{ll90}{An optional vector of the lower limits of the 90 percent confidence interval of the mean outcome in groups at time points.}

\item{ul90}{An optional vector of the upper limits of the 90 percent confidence interval of the mean outcome in groups at time points.}

\item{ll99}{An optional vector of the lower limits of the 99 percent confidence interval of the mean outcome in groups at time points.}

\item{ul99}{An optional vector of the upper limits of the 99 percent confidence interval of the mean outcome in groups at time points.}

\item{a}{An optional vector of minima of the outcome in groups at time points.}

\item{b}{An optional vector of maxima of the outcome in groups at time points.}

\item{lq}{An optional vector of lower quartils of the outcome in groups at time points.}

\item{uq}{An optional vector of upper quartils of the outcome in groups at time points.}

\item{data}{An optional data frame containing the study information.}

\item{meanSd.meanAB}{An optional character string indicating which method is used for obtaining mean and standard deviation
from mean and range. Either "range" for the range method (default), or "walterYao" for the method by Walter and Yao.}

\item{meanSd.mIqrRange}{An optional character string indicating which method is used for obtaining mean and standard deviation
from median, interquartile range, and range. Either "bland" for the method by bland, or "wanEtal" for the method by Wan et al. (default).}

\item{meanSd.mIqr}{An optional character string indicating which method is used for obtaining mean and standard deviation
from median and interquartile range. Either "wanEtal" for the method by Wan et al. (default), or "cochrane" for Cochrane method.}

\item{meanSd.mRange}{An optional character string indicating which method is used for obtaining mean and standard deviation
from median and range. Either "hozoDjulbegovicHozo" for the method by Hozo, Djulbegovic, and Hozo, or "wanEtal" for the method by Wan et al. (default).}

\item{meanSd.mean}{An optional character string indicating which method is used for obtaining mean and standard deviation
from mean and missing standard deviation by imputation. Either "furukawaEtal" for the method by Furukawa et al. (default), or "marinhoEtal" for the method by Marinho et al..}
}
\value{
A data frame object that contains study labels, group labels, time points of outcome measurement,
numbers of observations in groups at time points, means of outcome in groups at time points, standard deviations of outcome in groups at time points.
}
\description{
Uniforms a differently or incompletely reported continuous outcome across studies included in a meta-analysis.
}
\examples{
data(data2)
coUniform<-co.co(data=data2)
}
